Move tests from wire-tests to correct modules
