+++ wire-java-generator/src/test/java/com/squareup/wire/java/JavaGeneratorTest.java
