+++ wire-kotlin-generator/build.gradle
@@ -16,6 +16,6 @@ dependencies {
   api deps.kotlin.stdlib.jdk8
   implementation deps.okio
   api deps.kotlinpoet
-  testImplementation deps.junit
-  testImplementation deps.assertj
+  testImplementation project(':wire-test-utils')
+  testImplementation deps.kotlin.test.junit
 }
