+++ wire-schema/build.gradle
@@ -15,4 +15,8 @@ dependencies {
   api deps.guava
   annotationProcessor deps.autovalue.compiler
   compileOnly deps.autovalue.annotations
+  testImplementation project(':wire-test-utils')
+  testImplementation deps.junit
+  testImplementation deps.assertj
+  testImplementation deps.jimfs
 }
