+++ wire-schema/src/test/java/com/squareup/wire/schema/IdentifierSetTest.java
