+++ wire-schema/src/test/java/com/squareup/wire/schema/OptionsTest.java
