+++ wire-schema/src/test/java/com/squareup/wire/schema/internal/parser/ServiceElementTest.java
